// src/unified-map/standard/overlay/types.ts
export type StandardOverlayKind = Extract<
  OverlayKind,
  "marker" | "popup" | "dom" | "polyline" | "polygon" | "circle"
>;

export interface StandardOverlayOptions extends OverlayOptions {
  visible?: boolean;
  zIndex?: number;
  minZoom?: number;
  maxZoom?: number;
  metadata?: Record<string, unknown>;
}

export interface AnchoredOverlayOptions extends StandardOverlayOptions {
  coordinate: LngLatLike;
  anchor?: OverlayAnchor;
  offset?: PixelOffsetLike;
}

export interface PathOverlayOptions extends StandardOverlayOptions {
  coordinates: readonly LngLatLike[];
}
// ...
export interface PolylineStyle {
  color?: string;
  width?: number;
  opacity?: number;
  dashArray?: readonly number[];
  lineCap?: "butt" | "round" | "square";
  lineJoin?: "miter" | "round" | "bevel";
}

export interface PolygonStyle extends PolylineStyle {
  fillColor?: string;
  fillOpacity?: number;
}

export interface CircleStyle extends PolygonStyle {}

export interface PolylineOverlayOptions extends PathOverlayOptions {
  style?: PolylineStyle;
  curve?: boolean;
}

export interface PolygonOverlayOptions extends PathOverlayOptions {
  style?: PolygonStyle;
}

export interface CircleOverlayOptions extends AnchoredOverlayOptions {
  radius: number;
  style?: CircleStyle;
}

export interface BaseStandardOverlayDefinition<
  TKind extends StandardOverlayKind,
  TOptions extends object,
> extends OverlayDefinition<TOptions> {
  kind: TKind;
}
// ...
export interface PolygonOverlayDefinition
  extends BaseStandardOverlayDefinition<"polygon", PolygonOverlayOptions> {}